requirements.txt
astroid==2.3.3
bleach==3.1.0
bottle==0.12.17
bottle-cors==0.1.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.41.0
docutils==0.15.2
idna==2.8
importlib-metadata==1.2.0
isort==4.3.21
keyring==20.0.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
oauthlib==3.1.0
Pillow==7.1.1
pkginfo==1.5.0.1
psycopg2==2.8.4
Pygments==2.5.2
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
pywin32-ctypes==0.2.0
readme-renderer==24.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
sentry-sdk==0.13.5
six==1.13.0
sqlparse==0.3.0
tqdm==4.36.1
twine==3.1.1
typed-ast==1.4.0
urllib3==1.25.7
webencodings==0.5.1
wrapt==1.11.2
zipp==0.6.0
